Moving RunnerException to the api package
