+++ ls-mandelbrot-core/src/main/java/de/lexasoft/mandelbrot/api/MandelbrotRunner.java
@@ -1,7 +1,5 @@
 package de.lexasoft.mandelbrot.api;
 
-import de.lexasoft.mandelbrot.MandelbrotRunnerException;
-
 public interface MandelbrotRunner {
 
 	/**
